pyproject.toml
[tool.poetry]
name = "broker-api"
version = "0.1.0"
description = ""
authors = ["Valentin Kim <valentink@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.1"
requests = "^2.31.0"
django-environ = "^0.11.2"
uvicorn = "^0.27.0"
sseclient = "^0.0.27"
celery = "^5.3.6"
redis = "^5.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-django = "^4.4.0"
pytest-cov = "^3.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
